admin rule permissions working
added isAdmin in User.php and Kernel.php
added middleware to web.php routes

+++ routes/web.php
@@ -41,8 +41,8 @@ Route::get('/Places' , [PlacesController::class, 'index'])->name('Places.Index')
 Route::get('/Posts' , [PostsController::class, 'index'])->name('Posts.Index')->middleware('isloggedin'); //Lai pieklutu valstu skatam
 Route::get('/Vehicle' , [VehicleController::class, 'index'])->name('Vehicle.Vehicle')->middleware('isloggedin'); //Lai pieklutu transportu skatam
 
-Route::get('/AddNewPlace' , [PlacesController::class, 'create'])->name('Places.create')->middleware('isloggedin'); //Lai pieklutu valstu izveidosanas skatam
-Route::get('/Create' , [CreateVehicleController::class, 'index'])->name('Vehicle.create')->middleware('isloggedin'); //Lai pieklutu transportu izveidosanas skatam
+Route::get('/AddNewPlace' , [PlacesController::class, 'create'])->name('Places.create')->middleware('isloggedin', 'Admin'); //Lai pieklutu valstu izveidosanas skatam
+Route::get('/Create' , [CreateVehicleController::class, 'index'])->name('Vehicle.create')->middleware('isloggedin', 'Admin');//Lai pieklutu transportu izveidosanas skatam
 Route::get('/AddNewPost' , [PostsController::class, 'create'])->name('Posts.create')->middleware('isloggedin');
 
 Route::post('/storepost',[PostsController::class, 'store'])->name('Posts.store')->middleware('isloggedin');
@@ -53,15 +53,15 @@ Route::delete('/{posts}/delete', [PostsController::class, 'destroy'])->name('Pos
 Route::patch('/{posts}/update', [PostsController::class, 'update'])->name('Posts.update')->middleware('isloggedin');
 
 Route::get('/{vehicles}/show', [ShowController::class, 'show'])->name('Vehicle.show')->middleware('isloggedin');
-Route::get('/{vehicles}/edit', [EditController::class, 'edit'])->name('Vehicle.edit')->middleware('isloggedin');
-Route::delete('/{vehicles}', [DeleteController::class, 'destroy'])->name('Vehicle.delete')->middleware('isloggedin');
-Route::patch('/{vehicles}', [UpdateController::class, 'update'])->name('Vehicle.update')->middleware('isloggedin');
+Route::get('/{vehicles}/edit', [EditController::class, 'edit'])->name('Vehicle.edit')->middleware('isloggedin', 'Admin');
+Route::delete('/{vehicles}', [DeleteController::class, 'destroy'])->name('Vehicle.delete')->middleware('isloggedin', 'Admin');
+Route::patch('/{vehicles}', [UpdateController::class, 'update'])->name('Vehicle.update')->middleware('isloggedin', 'Admin');
 
 Route::post('/store',[StorePlacesController::class, 'store'])->name('Places.store')->middleware('isloggedin');
 Route::get('/{places}/show/this', [PlacesController::class, 'show'])->name('Places.show')->middleware('isloggedin');
-Route::get('/{places}/editing/this', [PlacesController::class, 'edit'])->name('Places.edit')->middleware('isloggedin');
-Route::delete('/{places}/delete/this', [PlacesController::class, 'destroy'])->name('Places.delete')->middleware('isloggedin');
-Route::patch('/{places}/update/this', [PlacesController::class, 'update'])->name('Places.update')->middleware('isloggedin');
+Route::get('/{places}/editing/this', [PlacesController::class, 'edit'])->name('Places.edit')->middleware('isloggedin', 'Admin');
+Route::delete('/{places}/delete/this', [PlacesController::class, 'destroy'])->name('Places.delete')->middleware('isloggedin', 'Admin');
+Route::patch('/{places}/update/this', [PlacesController::class, 'update'])->name('Places.update')->middleware('isloggedin', 'Admin');
 
 
 Route::get('/user', [IndexController::class, 'index'])->name('User.Index')->middleware('isloggedin');
